requirements.txt
gunicorn
Flask
tensorflow
google.cloud.storage
numpy
keras
tarfile
os